+++ retailer/src/main/java/com/generic/retailer/payload/request/ProductRequest.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 @Builder
 public class ProductRequest {
-    private long productId;
+    private int productId;
     private String name;
     private double price;
-    private long quantity;
+    private int quantity;
 }
